Disable auth on frontend to enable earlier experimentation.
This was motivated by a bug in the original file's google-oauth.js
implementation. On all of Firefox, Safari and Chrome, I was unable to
successfully authenticate. It would seem that when a successful google
authentication callback returned, the subsequent request's Cookie header
would not be set to the new authenticated cookie set in the callback.

This causes an indefinite redirect, preventing access to the Kibana
dashboard.